Refactor: Split ChatbotWidget into smaller components
This commit refactors the ChatbotWidget component into smaller, more manageable files. It aims to improve code organization and maintainability without altering the functionality of the chatbot. Unused imports and files are also removed.
