+++ src/hooks/chat/useChatState.ts
@@ -1,0 +1,42 @@
+
+import { useState, useRef } from 'react';
+
+export const useChatState = () => {
+  const [conversationStage, setConversationStage] = useState<"intro" | "questions" | "completion">("intro");
+  const [isResuming, setIsResuming] = useState(false);
+  const [currentSectionIndex, setCurrentSectionIndex] = useState(0);
+  const [currentQuestionIndex, setCurrentQuestionIndex] = useState(0);
+  const [formData, setFormData] = useState<Record<string, any>>({});
+  const [input, setInput] = useState("");
+  const [showOptions, setShowOptions] = useState(true);
+  const chatInitializedRef = useRef(false);
+  
+  const resetChatState = () => {
+    setInput("");
+    setShowOptions(true);
+    setConversationStage("intro");
+    setIsResuming(false);
+    setCurrentSectionIndex(0);
+    setCurrentQuestionIndex(0);
+    setFormData({});
+  };
+
+  return {
+    conversationStage,
+    setConversationStage,
+    isResuming,
+    setIsResuming,
+    currentSectionIndex,
+    setCurrentSectionIndex,
+    currentQuestionIndex,
+    setCurrentQuestionIndex,
+    formData,
+    setFormData,
+    input,
+    setInput,
+    showOptions,
+    setShowOptions,
+    chatInitializedRef,
+    resetChatState
+  };
+};
